Remove `TriggerableMonitor`
